pyproject.toml
[tool.poetry]
name = "realty-bot"
version = "0.1.0"
description = ""
authors = ["Alchemist of the web <nikitakzntcv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-debug-toolbar = "^4.4.6"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
celery = {extras = ["redis"], version = "^5.4.0"}
psycopg2-binary = "^2.9.9"
django-celery-beat = "^2.7.0"
pytz = "^2024.2"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
